+++ frontend/simpler/src/pages/verify-mps/MpsVerificationResults.tsx
@@ -1,12 +1,12 @@
-import MainHeader from "../components/ui/MainHeader.tsx";
-import MainNavBar from "../components/layout/MainNavBar.tsx";
-import PageContentHeader from "../components/ui/PageContentHeader.tsx";
-import BottomNavBar from "../components/layout/BottomNavBar.tsx";
+import MainHeader from "../../components/ui/MainHeader.tsx";
+import MainNavBar from "../../components/layout/MainNavBar.tsx";
+import PageContentHeader from "../../components/ui/PageContentHeader.tsx";
+import BottomNavBar from "../../components/layout/BottomNavBar.tsx";
 import {useNavigate, useParams} from "react-router-dom";
-import {MPS_DATA_SS_PREFIX, MPS_VERIF_SS_PREFIX} from "../utils/storageConstants.ts";
-import type {MpsVerificationResponse} from "../api/verification/verificationTypes.ts";
-import MpsVerificationInput from "../components/layout/mps/MpsVerificationInput.tsx";
-import MpsVerificationError from "../components/layout/mps/MpsVerificationError.tsx";
+import {MPS_DATA_SS_PREFIX, MPS_VERIF_SS_PREFIX} from "../../utils/storageConstants.ts";
+import type {MpsVerificationResponse} from "../../api/verification/verificationTypes.ts";
+import MpsVerificationInput from "../../components/layout/mps/MpsVerificationInput.tsx";
+import MpsVerificationError from "../../components/layout/mps/MpsVerificationError.tsx";
 import {get} from "idb-keyval";
 import {useEffect, useState} from "react";
