+++ frontend/simpler/src/pages/verify-mps/VerifyMpsLayout.tsx
@@ -1,0 +1,7 @@
+import {Outlet} from "react-router-dom";
+
+function VerifyMpsLayout() {
+    return (<><Outlet/></>)
+}
+
+export default VerifyMpsLayout
